[flow]: update operation with validation
